Add files via upload
This wellness bot is a small language model which could be converted to a tool augmented large language model. Due to constraint of time, I was only able to create small language model. This model is based on benchmark slm-eliza. This is the application of using language model for wellness of a person or psychotherapy. It leverages language model toolkit to interact with the user ascertain their mental health and provide feedback based on a corpus including psychotherapy text and papers. It would be similar to scholar agent in a way that it would read and collect data but would differentiate on the methods of processing the data and the output it provides.